requirements.txt
aiodns==3.0.0
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
aioredis==1.3.1
aiosqlite==0.16.1
amqp==5.0.6
appnope==0.1.2
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.16.29
botocore==1.19.63
cchardet==2.1.7
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cli-helpers==1.2.1
click==7.1.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
configobj==5.0.6
dataclasses==0.6
decorator==5.0.9
discord.py==1.7.3
dj-database-url==0.5.0
Django==3.2.5
elasticsearch==7.13.1
elasticsearch-dsl==7.4.0
fake-useragent==0.1.11
fancycompleter==0.9.1
FastAPI-SQLAlchemy==0.2.1
geographiclib==1.50
geopy==2.1.0
googletransx==2.4.2
greenlet==1.1.0
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
iso8601==0.1.14
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.1
odfe-sql-cli==1.13.0.0
pandas==1.2.4
parso==0.8.2
pdbpp==0.10.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.6
psutil==5.8.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycares==4.0.0
pycparser==2.20
pydantic==1.8.2
pyfiglet==0.8.post1
Pygments==2.6.1
pypika-tortoise==0.1.1
pyrepl==0.9.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-multipart==0.0.5
python-socks==1.2.4
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
-e git+https://github.com/xaviermathew/Rainman.git@0bc24336f5da1c7d1e1a37441396c9feec16aa0a#egg=rainman
requests==2.25.1
requests-aws4auth==0.9
requests-oauthlib==1.3.0
retry==0.9.2
s3transfer==0.3.7
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.18
SQLAlchemy-Utils==0.37.7
sqlparse==0.4.1
starlette==0.14.2
tabulate==0.8.9
terminaltables==3.1.0
tortoise-orm==0.17.4
tqdm==4.61.1
traitlets==5.0.5
tweepy @ git+https://github.com/tweepy/tweepy.git@145d7cfb902ed7bd10143e201ea89af41ba7d628
twint @ git+https://github.com/twintproject/twint.git@e7c8a0c764f6879188e5c21e25fb6f1f856a7221
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
wmctrl==0.4
yarl==1.6.3
